review: address PR feedback; break import cycle; drop typeclass framing
User-facing changes:
- Drop notes/operator-typeclass.md and the §6 categorical-framing section in
  docs/createOperator.md. The "duality" claim between operator and combinator
  forms was a stretch — the table and prose now describe them as two ways of
  getting to the same N-children-then-layout shape rather than duals.
- Drop ChartBuilder.facet() and .stack() (sugar for .flow(spread/stack)).
  One callsite — BarStackedFluent.stories.tsx — migrated to .flow(...).
  Their sole purpose was importing v3 spread/stack into chartBuilder.ts,
  which was the deepest source of the import cycle.

Cycle removed:
- The chartBuilder ⇄ createOperator cycle is gone. resolveMarkResult and
  LayerContext now live in chartBuilder.ts; createOperator.ts imports them
  one-way and re-exports for back-compat.
- v3 group is extracted from frame.tsx into graphicalOperators/group.ts so
  chartBuilder importing Frame doesn't transitively pull in createOperator.
- Bottom-of-file `import { createOperator }` workarounds in spread/scatter/
  table.tsx are no longer load-bearing — moved to top-of-file.

Code review fixes:
- #2: NameableMark consolidated to one definition in withGoFish.ts.
  createOperator's variant (which had .render()) is gone — no callsites
  actually used it; the misleading `base(undefined as any)` is removed.
- #3: connect.tsx:73 comment "center-to-center" → "center" after the rename.
- #5: SplitResult bucket type widened to `Datum | Datum[]` to match the
  no-`by` per-item path. Downstream `applyChannels` already normalised both
  forms, but the type now reflects what split actually returns.
- #7: Python group() takes `by: str` (no default) so missing-arg errors come
  from Python itself instead of a manual ValueError.
- Nit: applyChannels comment notes that the array short-circuit covers both
  combinator-form pre-built arrays and entry-flagged user-supplied arrays.

Merge-fallout fix:
- Stories using `Layer([Chart, Chart]).render(...)` reverted from lowercase
  `layer` to uppercase `Layer`. Lowercase `layer` is now the v3 mark form
  introduced by main's bottle PR (#366), which doesn't accept ChartBuilders
  or expose .render(). The chart-Array variant is `Layer` (uppercase).

Verified: build passes, widget bundle builds, 77/77 Python tests pass,
madge shows no chartBuilder/createOperator cycle outside of the lib.ts
barrel re-exports.
