+++ packages/gofish-graphics/stories/forwardsyntax/Bar/BarStackedFluent.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/html";
 import { initializeContainer } from "../../helper";
 import { seafood } from "../../../src/data/catch";
-import { Chart, rect } from "../../../src/lib";
+import { Chart, rect, spread, stack } from "../../../src/lib";
 
 const meta: Meta = {
   title: "Forward Syntax V3/Bar/Stacked Fluent",
@@ -18,15 +18,16 @@ export default meta;
 
 type Args = { w: number; h: number };
 
-// Uses new .facet() and .stack() methods instead of .flow(spread(...), stack(...))
 export const Default: StoryObj<Args> = {
   args: { w: 400, h: 400 },
   render: (args: Args) => {
     const container = initializeContainer();
 
     Chart(seafood)
-      .facet({ by: "lake",  dir: "x" })
-      .stack({ by: "species",  dir: "y" })
+      .flow(
+        spread({ by: "lake", dir: "x" }),
+        stack({ by: "species", dir: "y" })
+      )
       .mark(rect({ h: "count", fill: "species" }))
       .render(container, { w: args.w, h: args.h, axes: true });
