+++ packages/gofish-graphics/stories/forwardsyntax/Streamgraph.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/html";
 import { initializeContainer } from "../helper";
 import { seafood } from "../../src/data/catch";
-import { Chart, spread, blank, stack, layer, select } from "../../src/lib";
+import { Chart, spread, blank, stack, Layer, select } from "../../src/lib";
 import { area, group } from "../../src/lib";
 
 const meta: Meta = {
@@ -24,7 +24,7 @@ export const Default: StoryObj<Args> = {
   render: (args: Args) => {
     const container = initializeContainer();
 
-    layer([
+    Layer([
       Chart(seafood)
         .flow(
           spread({ by: "lake",  dir: "x", spacing: 64, alignment: "middle" }),
