+++ packages/gofish-graphics/stories/vega-lite/ColoredScatterPlot.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/html";
 import { initializeContainer } from "../helper";
-import { Chart, layer, scatter, circle } from "../../src/lib";
+import { Chart, Layer, scatter, circle } from "../../src/lib";
 import data from "vega-datasets";
 const meta: Meta = {
   title: "Vega-Lite/Colored Scatter Plot",
@@ -37,7 +37,7 @@ export const Default: StoryObj<Args> = {
       const speciesList = Array.from(new Set(penguins.map((d) => d["Species"])));
       const bySpecies = (species: string) => penguins.filter((d) => d["Species"] === species);
 
-    layer(
+    Layer(
       speciesList.map((species) => () =>
         Chart(bySpecies(species))
           .flow(
